Changes the location of component
